Update of Bluewing Client to build 96, Server to build 28
Updated both to DarkEdif v11.

Client:
• Updated to DarkEdif SDK v11.
• Now compatible with Android (beta; Android 4.4+) and iOS (alpha; iOS 10.0+)
• Fixed crash on channel peer named loops and public channel listing named loops

Bluewing Server:
• Updated to DarkEdif SDK v11.
• Added an error check when name is blank in a loop all clients named loop, matching the other loops.

Code cleanup:
• Changed types to lw_XX types, and some native-size to fixed-size.
• Hid most warnings on Clang during Android/iOS compilations
• Prevented filename repeats due to Clang confusion, even filenames in different folders; notably, src/unix/common.h was renamed src/unix/unix common.h
• Removed extraneous Lacewing/include/lacewing.h file, as there is already Lacewing/Lacewing.h file with much the same content.

+++ DarkEdif/Lib/Shared/Lacewing/CodePointAllowList.cpp
@@ -131,7 +131,7 @@ std::string lacewing::codepointsallowlist::setcodepointsallowedlist(std::string 
 				if (std::find(specificCodePoints.cbegin(), specificCodePoints.cend(), codePointAllowed) != specificCodePoints.cend())
 					return CPALMakeError(this, acTemp, "Specific codepoint %lu was added twice in list \"%hs\".", codePointAllowed, acStr.c_str());
 
-				specificCodePoints.push_back(codePointAllowed);
+				specificCodePoints.push_back((int)codePointAllowed);
 				if (cur[0] == ',')
 					++cur;
 				goto nextChar;
